Moved many things around, extracted DB string into env variable

+++ src/ferdabot/ferdabot.go
@@ -5,7 +5,6 @@ import (
 	"os"
 	"os/signal"
 	"regexp"
-	"strings"
 	"syscall"
 	"time"
 
@@ -42,9 +41,10 @@ func NewBot() Bot {
 }
 
 func (b *Bot) Setup() error {
+	// region Discord Conn
 	token := os.Getenv("FERDATOKEN")
 	if token == "" {
-		fmt.Println("A bot environment token is required. FERDATOKEN was empty.")
+		return fmt.Errorf("A bot environment token is required. FERDATOKEN was empty.\n")
 	}
 
 	dg, err := discordgo.New("Bot " + token)
@@ -52,22 +52,32 @@ func (b *Bot) Setup() error {
 		return err
 	}
 
-	// region config
 	dg.AddHandler(b.messageCreate)
 	dg.Identify.Intents = discordgo.MakeIntent(discordgo.IntentsGuildMessages)
 	// endregion
 
-	dbCon, sqlErr := sqlx.Connect("postgres", "host=db user=postgres dbname=postgres password=pass sslmode=disable")
-	if sqlErr != nil {
-		fmt.Printf("Error connecting to DB: %s\n", sqlErr)
+	// region DB
+	connURL := os.Getenv("DBCONNURL")
+	if connURL == "" {
+		return fmt.Errorf("A db connection string is required. DBCONNURL was empty.\n")
 	}
 
-	b.db = dbCon
-	b.dg = dg
+	dbCon, sqlErr := sqlx.Connect("postgres", connURL)
+	if sqlErr != nil {
+		return fmt.Errorf("Error connecting to DB: %s\n", sqlErr)
+	}
+	// endregion
 
+	// region OS Signal Channel
 	sc := make(chan os.Signal, 1)
 	signal.Notify(sc, os.Interrupt, syscall.SIGINT, syscall.SIGTERM, os.Kill)
+	// endregion
+
+	// region Assignments
+	b.db = dbCon
+	b.dg = dg
 	b.signalChannel = sc
+	// endregion
 
 	return nil
 }
@@ -76,8 +86,7 @@ func (b *Bot) Run() error {
 	// region connecting
 	conErr := b.dg.Open()
 	if conErr != nil {
-		fmt.Println("Couldn't open discord connection.")
-		os.Exit(2)
+		return fmt.Errorf("Couldn't open discord connection, %s\n", conErr)
 	}
 
 	fmt.Println("Startup (ok)")
@@ -88,116 +97,4 @@ func (b *Bot) Run() error {
 	}
 
 	return nil
-}
-
-func (b *Bot) messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-	m.Content = strings.TrimSpace(m.Content)
-	// Ignore the message if the bot sent it
-	if m.Author.ID == s.State.User.ID {
-		return
-	}
-
-	b.handleCommands(s, m)
-
-}
-
-func (b *Bot) handleCommands(s *discordgo.Session, m *discordgo.MessageCreate) {
-	fmt.Printf("Message %s\n", m.Content)
-	splitMessage := strings.Split(m.Content, " ")
-	command := splitMessage[0]
-	data := strings.Join(splitMessage[1:], " ")
-	switch command {
-	case "!echo":
-		b.handleEcho(s, m, data)
-	case "+ferda":
-		b.handleNewFerda(s, m, data)
-	case "?ferda":
-		b.handleGetFerda(s, m, data)
-	case "?help", "!help", "+help":
-		if _, err := s.ChannelMessageSend(m.ChannelID, "Use `+ferda @User [reason]` to add a ferda, and `?ferda @User` to get a ferda."); err != nil {
-			fmt.Printf("Error sending message to discord %s\n", err)
-		}
-	}
-}
-
-func (b *Bot) handleEcho(s *discordgo.Session, m *discordgo.MessageCreate, trimmedText string) {
-	for _, phrase := range BannedEchoPhrases {
-		if strings.Contains(trimmedText, phrase) {
-			if _, err := s.ChannelMessageSend(m.ChannelID, "You shouldn't be trying to send that you dumb fuck."); err != nil {
-				fmt.Printf("Error occured responding to ping. %s\n", err)
-			}
-			return
-		}
-	}
-
-	if _, err := s.ChannelMessageSend(m.ChannelID, trimmedText); err != nil {
-		fmt.Printf("Error occured responding to ping. %s\n", err)
-	}
-}
-
-func (b *Bot) getUserFromText(trimmedText string) string {
-	split := strings.Split(trimmedText, " ")
-	user := split[0]
-
-	found := userRegex.Find([]byte(user))
-	foundString := string(found)
-	return foundString
-}
-
-func (b *Bot) handleNewFerda(s *discordgo.Session, m *discordgo.MessageCreate, trimmedText string) {
-	split := strings.Split(trimmedText, " ")
-	foundString := b.getUserFromText(trimmedText)
-
-	if foundString == "" {
-		if _, err := s.ChannelMessageSend(m.ChannelID, "You must ping someone who is ferda. Ex: `!ferda @Logan is a great guy`"); err != nil {
-			fmt.Printf("Error occured responding to ping. %s\n", err)
-		}
-		return
-	}
-
-	res, dbErr := b.db.NamedExec(`INSERT INTO ferda (userid, time, reason, creatorid) VALUES (:userid, :time, :reason, :creatorid)`, map[string]interface{}{
-		"userid":    foundString,
-		"time":      time.Now(),
-		"reason":    strings.Join(split[1:], " "),
-		"creatorid": m.Author.ID,
-	})
-	if dbErr != nil {
-		fmt.Printf("Error inserting into the DB %s\n", dbErr)
-	}
-	count, _ := res.RowsAffected()
-	if count == 0 {
-		if _, err := s.ChannelMessageSend(m.ChannelID, "No rows effected."); err != nil {
-			fmt.Printf("Error sending a message to discord.")
-			return
-		}
-	}
-
-	if _, err := s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("Thanks for the new ferda <@!%s>", m.Author.ID)); err != nil {
-		fmt.Printf("Error sending a message to discord.")
-		return
-	}
-}
-
-func (b *Bot) handleGetFerda(s *discordgo.Session, m *discordgo.MessageCreate, trimmedText string) {
-	foundUser := b.getUserFromText(trimmedText)
-	user, err := s.User(foundUser)
-	if err != nil {
-		fmt.Printf("Couldn't load user from discord: %s\n", err)
-	}
-
-	ferdaEntry := FerdaEntry{}
-	dbErr := b.db.Get(&ferdaEntry, `SELECT * FROM ferda WHERE userid = $1 ORDER BY RANDOM()`, foundUser)
-	if dbErr != nil {
-		if dbErr.Error() == "sql: no rows in result set" {
-			if _, err := s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("<@!%s> is not ferda.", user.ID)); err != nil {
-				fmt.Printf("Error sending message to discord %s\n", err)
-			}
-		}
-		fmt.Printf("Error selecting from table %+v\n", dbErr)
-		return
-	}
-
-	if _, err := s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("<@!%s> was ferda on %s for %s", user.ID, ferdaEntry.When.Format("Mon, Jan _2 2006"), ferdaEntry.Reason)); err != nil {
-		fmt.Printf("Error sending message to discord %s\n", err)
-	}
 }
